grate: factor out 2d related functions to grate-2d.c
Logical split, no functional change.
